build: 项目结构优化
