+++ demo/src/main/java/com/tlcsdm/demo/FXMenubarConfigration.java
@@ -27,15 +27,8 @@
 
 package com.tlcsdm.demo;
 
-import com.tlcsdm.frame.MenubarConfigration;
-
-import javafx.scene.control.CheckMenuItem;
-import javafx.scene.control.Menu;
-import javafx.scene.control.MenuBar;
-import javafx.scene.control.MenuItem;
-import javafx.scene.control.RadioMenuItem;
-import javafx.scene.control.SeparatorMenuItem;
-import javafx.scene.control.ToggleGroup;
+import com.tlcsdm.frame.service.MenubarConfigration;
+import javafx.scene.control.*;
 
 public class FXMenubarConfigration implements MenubarConfigration {
