--- demo/src/main/resources/META-INF/services/com.tlcsdm.login.LoginCheck
@@ -1,1 +1,0 @@
-com.tlcsdm.demo.ControlsLoginCheck
