+++ frame/src/main/java/com/tlcsdm/frame/FXSampler.java
@@ -42,7 +42,7 @@ import com.tlcsdm.core.javafx.util.JavaFxSystemUtil;
 import com.tlcsdm.core.javafx.util.StageUtils;
 import com.tlcsdm.core.util.InterfaceScanner;
 import com.tlcsdm.frame.model.*;
-import com.tlcsdm.frame.service.SamplePostProcessorService;
+import com.tlcsdm.frame.service.*;
 import com.tlcsdm.frame.util.I18nUtils;
 import com.tlcsdm.frame.util.SampleScanner;
 import javafx.application.Application;
