--- frame/src/main/java/com/tlcsdm/frame/SplashScreen.java
@@ -1,14 +1,0 @@
-package com.tlcsdm.frame;
-
-import javafx.scene.image.Image;
-
-/**
- * 闪屏图片
- *
- * @author: unknowIfGuestInDream
- * @date: 2023/3/3 22:41
- */
-public interface SplashScreen {
-
-    Image getImage();
-}
