+++ frame/src/main/java/com/tlcsdm/frame/service/CenterPanelService.java
@@ -25,8 +25,9 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.tlcsdm.frame;
+package com.tlcsdm.frame.service;
 
+import com.tlcsdm.frame.Sample;
 import com.tlcsdm.frame.model.Project;
 import com.tlcsdm.frame.model.WelcomePage;
 import javafx.scene.Node;
