+++ frame/src/main/java/com/tlcsdm/frame/service/FXSamplerConfiguration.java
@@ -24,30 +24,20 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package com.tlcsdm.frame.service;
 
-package com.tlcsdm.frame;
+import javafx.scene.image.Image;
 
-import javafx.scene.control.MenuBar;
-import javafx.stage.Stage;
+public interface FXSamplerConfiguration {
+    String getSceneStylesheet();
 
-/**
- * 菜单栏配置(应用模块实现)
- */
-public interface MenubarConfigration {
+    /**
+     * UI的名称
+     */
+    String getStageTitle();
 
-    MenuBar menuBar = new MenuBar();
-
-    MenuBar setMenuBar(MenuBar menuBar);
-
-    default MenuBar getMenuBar() {
-        Stage primaryStage = FXSampler.getStage();
-        setMenuBar(menuBar);
-        // 设置菜单条长度
-        menuBar.setPrefWidth(primaryStage.getWidth());
-        // 宽度监听设置菜单条长度
-        menuBar.widthProperty()
-                .addListener((observable, oldValue, newValue) -> menuBar.setPrefWidth(newValue.doubleValue()));
-        return menuBar;
-    }
-
+    /**
+     * 程序logo
+     */
+    Image getAppIcon();
 }
