+++ login/src/main/java/module-info.java
@@ -5,6 +5,8 @@ module com.tlcsdm.login {
 
     opens com.tlcsdm.login to javafx.fxml;
     exports com.tlcsdm.login;
+    exports com.tlcsdm.login.service;
+    opens com.tlcsdm.login.service to javafx.fxml;
 
-    uses com.tlcsdm.login.LoginCheck;
+    uses com.tlcsdm.login.service.LoginCheck;
 }
