+++ smc/src/main/java/module-info.java
@@ -1,3 +1,34 @@
+/*
+ * Copyright (c) 2019, 2023 unknowIfGuestInDream
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *     * Redistributions of source code must retain the above copyright
+ * notice, this list of conditions and the following disclaimer.
+ *     * Redistributions in binary form must reproduce the above copyright
+ * notice, this list of conditions and the following disclaimer in the
+ * documentation and/or other materials provided with the distribution.
+ *     * Neither the name of unknowIfGuestInDream, any associated website, nor the
+ * names of its contributors may be used to endorse or promote products
+ * derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL UNKNOWIFGUESTINDREAM BE LIABLE FOR ANY
+ * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+import com.tlcsdm.frame.service.*;
+import com.tlcsdm.login.service.LoginCheck;
+import com.tlcsdm.smc.provider.*;
+
 module com.tlcsdm.smc {
     requires java.desktop;
     requires java.net.http;
@@ -22,20 +53,20 @@ module com.tlcsdm.smc {
     exports com.tlcsdm.smc.codeDev to com.tlcsdm.frame;
     exports com.tlcsdm.smc.codeDev.ecm to com.tlcsdm.frame;
     exports com.tlcsdm.smc.unitDesign to com.tlcsdm.frame;
-    exports com.tlcsdm.smc.provider to com.tlcsdm.frame, com.tlcsdm.login;
     exports com.tlcsdm.smc.config to com.tlcsdm.core;
 
     opens com.tlcsdm.smc.codeDev to javafx.graphics;
     opens com.tlcsdm.smc.tools to javafx.graphics;
     opens com.tlcsdm.smc.unitDesign to javafx.graphics;
     opens com.tlcsdm.smc.unitTest to javafx.graphics;
+    exports com.tlcsdm.smc.provider;
 
-    provides com.tlcsdm.frame.FXSamplerProject with com.tlcsdm.smc.SmcSamplerProject;
-    provides com.tlcsdm.frame.MenubarConfigration with com.tlcsdm.smc.SmcMenubarConfigration;
-    provides com.tlcsdm.frame.CenterPanelService with com.tlcsdm.smc.SmcCenterPanelService;
-    provides com.tlcsdm.frame.FXSamplerConfiguration with com.tlcsdm.smc.SmcConfiguration;
-    provides com.tlcsdm.frame.SplashScreen with com.tlcsdm.smc.SmcSplash;
-    provides com.tlcsdm.login.LoginCheck with com.tlcsdm.smc.SmcLoginCheck;
+    provides FXSamplerProject with SmcSamplerProjectProvider;
+    provides MenubarConfigration with SmcMenubarConfigrationProvider;
+    provides CenterPanelService with SmcCenterPanelProvider;
+    provides FXSamplerConfiguration with SmcConfigurationProvider;
+    provides SplashScreen with SmcSplashProvider;
+    provides LoginCheck with SmcLoginCheckProvider;
     provides com.tlcsdm.core.freemarker.TemplateLoaderService with com.tlcsdm.smc.config.SmcTemplateLoaderProvider;
     provides com.tlcsdm.frame.service.SamplePostProcessorService
             with com.tlcsdm.smc.provider.SmcSamplePostProcessorProvider;
