--- smc/src/main/resources/META-INF/services/com.tlcsdm.frame.FXSamplerConfiguration
@@ -1,1 +1,0 @@
-com.tlcsdm.smc.SmcConfiguration
