--- smc/src/main/resources/META-INF/services/com.tlcsdm.frame.MenubarConfigration
@@ -1,1 +1,0 @@
-com.tlcsdm.smc.SmcMenubarConfigration
